Add (book) management features and UI components:
This commit introduces a comprehensive set of features for adding, editing, and displaying books within the application. Key additions include:

- A new `AddBook` screen for users to input book details and upload images.
- A `BookModel` class to represent book data.
- UI components for displaying book images, titles, authors, and other details.
- Functionality for fetching categories and order types using BLoC.
- Integration with Supabase for data storage and retrieval.
- A dedicated `EditBook` screen for modifying existing book entries.
- Removal of the old `book_item` page, consolidating book management into a more streamlined interface.

These enhancements improve the user experience for managing books in the application.
