+++ lib/screens/book/book_item/book_item.dart
@@ -1,5 +1,5 @@
 import 'package:biblio/cubit/favorite_function/my_favorite_books_list/my_list_cubit.dart';
-import 'package:biblio/screens/book_item/book_page.dart';
+import 'package:biblio/screens/book/book_page/book_page.dart';
 import 'package:biblio/utils/components/border_radius.dart';
 import 'package:biblio/utils/components/height.dart';
 import 'package:biblio/utils/constants/colors_constants.dart';
